Fix typo
query interval
