+++ src/airflow/airqo_etl_utils/constants.py
@@ -30,6 +30,14 @@ class Frequency(Enum):
             return ""
 
 
+class DataSource(Enum):
+    THINGSPEAK = 1
+    CLARITY = 2
+    PLUME_LABS = 3
+    AIRNOW = 4
+    TAHMO = 5
+
+
 class Tenant(Enum):
     NASA = 1
     URBAN_BETTER = 2
